+++ python_packages/jupyter_lsp/jupyter_lsp/tests/test_bad_spec.py
@@ -1,7 +1,8 @@
 import pytest
 import traitlets
 
-from jupyter_lsp.session import LanguageServerSession
+from jupyter_lsp.schema import SPEC_VERSION
+from jupyter_lsp.session import LanguageServerSessionStdio
 
 
 @pytest.mark.parametrize(
@@ -12,8 +13,14 @@ from jupyter_lsp.session import LanguageServerSession
         {"languages": None},
         {"languages": 1},
         {"languages": [1, "two"]},
+        {
+            "argv": ["command"],
+            "languages": ["some language"],
+            "version": SPEC_VERSION,
+            "mode": "unknown",
+        },
     ],
 )
 def test_bad_spec(spec):
     with pytest.raises(traitlets.TraitError):
-        LanguageServerSession(spec=spec)
+        LanguageServerSessionStdio(spec=spec)